+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/execution/deprecated/cairo0_execution.rs
@@ -14,11 +14,13 @@ use blockifier::execution::deprecated_entry_point_execution::{
 use blockifier::execution::entry_point::{EntryPointExecutionContext, ExecutableCallEntryPoint};
 use blockifier::execution::errors::EntryPointExecutionError;
 use blockifier::execution::execution_utils::Args;
+use blockifier::execution::syscalls::vm_syscall_utils::SyscallUsageMap;
 use blockifier::state::state_api::State;
 use cairo_vm::hint_processor::hint_processor_definition::HintProcessor;
 use cairo_vm::vm::runners::cairo_runner::{CairoArg, CairoRunner};
 
 // blockifier/src/execution/deprecated_execution.rs:36 (execute_entry_point_call)
+#[expect(clippy::result_large_err)]
 pub(crate) fn execute_entry_point_call_cairo0(
     call: ExecutableCallEntryPoint,
     compiled_class_v0: CompiledClassV0,
@@ -82,13 +84,15 @@ pub(crate) fn execute_entry_point_call_cairo0(
 
     Ok(CallInfoWithExecutionData {
         call_info: execution_result,
-        syscall_usage,
+        syscall_usage_vm_resources: syscall_usage,
+        syscall_usage_sierra_gas: SyscallUsageMap::default(),
         vm_trace: None,
     })
     // endregion
 }
 
 // blockifier/src/execution/deprecated_execution.rs:192 (run_entry_point)
+#[expect(clippy::result_large_err)]
 pub fn cheatable_run_entry_point(
     runner: &mut CairoRunner,
     hint_processor: &mut dyn HintProcessor,