+++ crates/cheatnet/src/runtime_extensions/cheatable_starknet_runtime_extension.rs
@@ -3,19 +3,20 @@ use crate::runtime_extensions::call_to_blockifier_runtime_extension::execution::
 };
 use crate::state::CheatnetState;
 use anyhow::Result;
+use blockifier::blockifier_versioned_constants::SyscallGasCost;
 use blockifier::execution::entry_point::EntryPointExecutionContext;
 use blockifier::execution::syscalls::hint_processor::OUT_OF_GAS_ERROR;
-use blockifier::execution::syscalls::{
-    SyscallRequest, SyscallRequestWrapper, SyscallResponse, SyscallResponseWrapper,
-    syscall_base::SyscallResult,
+use blockifier::execution::syscalls::syscall_base::SyscallResult;
+use blockifier::execution::syscalls::syscall_executor::SyscallExecutor;
+use blockifier::execution::syscalls::vm_syscall_utils::{
+    RevertData, SyscallRequest, SyscallRequestWrapper, SyscallResponse, SyscallResponseWrapper,
+    SyscallSelector,
 };
 use blockifier::execution::{
     common_hints::HintExecutionResult,
-    deprecated_syscalls::DeprecatedSyscallSelector,
     syscalls::hint_processor::{SyscallExecutionError, SyscallHintProcessor},
 };
 use blockifier::utils::u64_from_usize;
-use blockifier::versioned_constants::SyscallGasCost;
 use cairo_vm::{
     types::relocatable::Relocatable,
     vm::{
@@ -26,8 +27,6 @@ use cairo_vm::{
 use conversions::string::TryFromHexStr;
 use runtime::{ExtendedRuntime, ExtensionLogic, StarknetRuntime, SyscallHandlingResult};
 use starknet_types_core::felt::Felt;
-
-pub type SyscallSelector = DeprecatedSyscallSelector;
 
 pub struct CheatableStarknetRuntimeExtension<'a> {
     pub cheatnet_state: &'a mut CheatnetState,
@@ -40,7 +39,7 @@ impl<'a> ExtensionLogic for CheatableStarknetRuntimeExtension<'a> {
 
     fn override_system_call(
         &mut self,
-        selector: DeprecatedSyscallSelector,
+        selector: SyscallSelector,
         vm: &mut VirtualMachine,
         extended_runtime: &mut Self::Runtime,
     ) -> Result<SyscallHandlingResult, HintError> {
@@ -88,13 +87,29 @@ impl<'a> ExtensionLogic for CheatableStarknetRuntimeExtension<'a> {
                     SyscallSelector::GetBlockHash,
                 )
                 .map(|()| SyscallHandlingResult::Handled),
+            SyscallSelector::StorageRead => self
+                .execute_syscall(
+                    syscall_handler,
+                    vm,
+                    cheated_syscalls::storage_read,
+                    SyscallSelector::StorageRead,
+                )
+                .map(|()| SyscallHandlingResult::Handled),
+            SyscallSelector::StorageWrite => self
+                .execute_syscall(
+                    syscall_handler,
+                    vm,
+                    cheated_syscalls::storage_write,
+                    SyscallSelector::StorageWrite,
+                )
+                .map(|()| SyscallHandlingResult::Handled),
             _ => Ok(SyscallHandlingResult::Forwarded),
         }
     }
 
     fn handle_system_call_signal(
         &mut self,
-        selector: DeprecatedSyscallSelector,
+        selector: SyscallSelector,
         _vm: &mut VirtualMachine,
         extended_runtime: &mut Self::Runtime,
     ) {
@@ -133,6 +148,8 @@ fn get_syscall_cost(
         SyscallSelector::Deploy => gas_costs.syscalls.deploy,
         SyscallSelector::GetExecutionInfo => gas_costs.syscalls.get_execution_info,
         SyscallSelector::GetBlockHash => gas_costs.syscalls.get_block_hash,
+        SyscallSelector::StorageRead => gas_costs.syscalls.storage_read,
+        SyscallSelector::StorageWrite => gas_costs.syscalls.storage_write,
         _ => unreachable!("Syscall has no associated cost"),
     }
 }
@@ -187,11 +204,11 @@ impl CheatableStarknetRuntimeExtension<'_> {
 
         if gas_counter < required_gas {
             //  Out of gas failure.
-            let out_of_gas_error = TryFromHexStr::try_from_hex_str(OUT_OF_GAS_ERROR)
-                .map_err(SyscallExecutionError::from)?;
+            let out_of_gas_error =
+                Felt::try_from_hex_str(OUT_OF_GAS_ERROR).map_err(SyscallExecutionError::from)?;
             let response: SyscallResponseWrapper<Response> = SyscallResponseWrapper::Failure {
                 gas_counter,
-                error_data: vec![out_of_gas_error],
+                revert_data: RevertData::new_normal(vec![out_of_gas_error]),
             };
             response.write(vm, &mut syscall_handler.syscall_ptr)?;
 
@@ -215,7 +232,7 @@ impl CheatableStarknetRuntimeExtension<'_> {
             Err(SyscallExecutionError::Revert { error_data: data }) => {
                 SyscallResponseWrapper::Failure {
                     gas_counter: remaining_gas,
-                    error_data: data,
+                    revert_data: RevertData::new_normal(data),
                 }
             }
             Err(error) => return Err(error.into()),