+++ crates/forge-runner/src/test_case_summary.rs
@@ -11,7 +11,6 @@ use cheatnet::runtime_extensions::call_to_blockifier_runtime_extension::rpc::Use
 use cheatnet::runtime_extensions::forge_runtime_extension::contracts_data::ContractsData;
 use conversions::byte_array::ByteArray;
 use conversions::felt::ToShortString;
-use debugging::CollectorError;
 use foundry_ui::UI;
 use num_traits::Pow;
 use shared::utils::build_readable_text;
@@ -82,13 +81,13 @@ impl GasStatistics {
         }
     }
 
-    #[allow(clippy::cast_precision_loss)]
+    #[expect(clippy::cast_precision_loss)]
     fn mean(gas_usages: &[u64]) -> f64 {
         let sum: f64 = gas_usages.iter().map(|&x| x as f64).sum();
         sum / gas_usages.len() as f64
     }
 
-    #[allow(clippy::cast_precision_loss)]
+    #[expect(clippy::cast_precision_loss)]
     fn std_deviation(mean: f64, gas_usages: &[u64]) -> f64 {
         let sum_squared_diff = gas_usages
             .iter()
@@ -136,7 +135,7 @@ pub enum TestCaseSummary<T: TestType> {
         /// Message to be printed after the test case run
         msg: Option<String>,
         /// Trace of the test case run
-        debugging_trace: Option<Result<debugging::Trace, CollectorError>>,
+        debugging_trace: Option<debugging::Trace>,
         /// Information on used gas
         gas_info: <T as TestType>::GasInfo,
         /// Resources used during test
@@ -153,7 +152,7 @@ pub enum TestCaseSummary<T: TestType> {
         /// Message returned by the test case run
         msg: Option<String>,
         /// Trace of the test case run
-        debugging_trace: Option<Result<debugging::Trace, CollectorError>>,
+        debugging_trace: Option<debugging::Trace>,
         /// Random arguments used in the fuzz test case run
         fuzzer_args: Vec<String>,
         /// Statistics of the test run
@@ -169,6 +168,8 @@ pub enum TestCaseSummary<T: TestType> {
 }
 
 #[derive(Debug)]
+// We allow large enum variant because `Single` is the bigger variant and it is used most often
+#[expect(clippy::large_enum_variant)]
 pub enum AnyTestCaseSummary {
     Fuzzing(TestCaseSummary<Fuzzing>),
     Single(TestCaseSummary<Single>),
@@ -195,7 +196,7 @@ impl<T: TestType> TestCaseSummary<T> {
     }
 
     #[must_use]
-    pub fn debugging_trace(&self) -> Option<&Result<debugging::Trace, CollectorError>> {
+    pub fn debugging_trace(&self) -> Option<&debugging::Trace> {
         match self {
             TestCaseSummary::Passed {
                 debugging_trace, ..
@@ -318,6 +319,7 @@ impl TestCaseSummary<Single> {
             used_resources,
             encountered_errors,
             fuzzer_args,
+            fork_data,
         }: RunCompleted,
         test_case: &TestCaseWithResolvedConfig,
         contracts_data: &ContractsData,
@@ -332,7 +334,7 @@ impl TestCaseSummary<Single> {
             contracts_data,
             trace_verbosity,
             name.clone(),
-            test_case.config.fork_config.clone(),
+            &fork_data,
         );
 
         match status {
@@ -347,6 +349,7 @@ impl TestCaseSummary<Single> {
                         trace_data: VersionedProfilerCallTrace::V1(build_profiler_call_trace(
                             &call_trace,
                             contracts_data,
+                            &fork_data,
                             versioned_program_path,
                         )),
                         debugging_trace,
@@ -384,6 +387,7 @@ impl TestCaseSummary<Single> {
                             trace_data: VersionedProfilerCallTrace::V1(build_profiler_call_trace(
                                 &call_trace,
                                 contracts_data,
+                                &fork_data,
                                 versioned_program_path,
                             )),
                             debugging_trace,
@@ -446,7 +450,7 @@ impl AnyTestCaseSummary {
     }
 
     #[must_use]
-    pub fn debugging_trace(&self) -> Option<&Result<debugging::Trace, CollectorError>> {
+    pub fn debugging_trace(&self) -> Option<&debugging::Trace> {
         match self {
             AnyTestCaseSummary::Fuzzing(case) => case.debugging_trace(),
             AnyTestCaseSummary::Single(case) => case.debugging_trace(),