+++ crates/sncast/tests/e2e/invoke.rs
@@ -14,7 +14,7 @@ use indoc::indoc;
 use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
 use snapbox::cmd::cargo_bin;
 use sncast::AccountType;
-use sncast::helpers::constants::{ARGENT_CLASS_HASH, BRAAVOS_CLASS_HASH, OZ_CLASS_HASH};
+use sncast::helpers::constants::{BRAAVOS_CLASS_HASH, OZ_CLASS_HASH, READY_CLASS_HASH};
 use sncast::helpers::fee::FeeArgs;
 use starknet::core::types::TransactionReceipt::Invoke;
 use starknet_types_core::felt::{Felt, NonZeroFelt};
@@ -61,7 +61,7 @@ async fn test_happy_case_human_readable() {
 
 #[test_case(DEVNET_OZ_CLASS_HASH_CAIRO_0.parse().unwrap(), AccountType::OpenZeppelin; "cairo_0_class_hash")]
 #[test_case(OZ_CLASS_HASH, AccountType::OpenZeppelin; "cairo_1_class_hash")]
-#[test_case(ARGENT_CLASS_HASH, AccountType::Argent; "argent_class_hash")]
+#[test_case(READY_CLASS_HASH, AccountType::Ready; "READY_CLASS_HASH")]
 #[test_case(BRAAVOS_CLASS_HASH, AccountType::Braavos; "braavos_class_hash")]
 #[tokio::test]
 async fn test_happy_case(class_hash: Felt, account_type: AccountType) {
@@ -219,7 +219,6 @@ fn test_wrong_function_name() {
     );
 }
 
-// TODO(#3116): Before, the error message included "Failed to deserialize param #2", but now it's an undecoded felt.
 #[test]
 fn test_wrong_calldata() {
     let args = vec![
@@ -245,7 +244,7 @@ fn test_wrong_calldata() {
         output,
         indoc! {r"
         Command: invoke
-        Error: Transaction execution error [..]0x4661696c656420746f20646573657269616c697a6520706172616d202332[..]
+        Error: Transaction execution error [..]Failed to deserialize param #2[..]
         "},
     );
 }